Support bi-directional stream
