+++ example/server.ts
@@ -1,4 +1,4 @@
-import { Server, Dispatcher } from "../server.ts";
+import { Session, Dispatcher } from "../session.ts";
 
 const hostname = "localhost";
 const port = 18800;
@@ -7,17 +7,27 @@ const dispatcher: Dispatcher = {
   async sum(x: number, y: number): Promise<number> {
     return x + y;
   },
+
+  async hello_server(name: string): Promise<string> {
+    return `Hello ${name}, this is server`;
+  },
+
+  async hello_client(name: string): Promise<string> {
+    // NOTE: 'this' is an instance of Session
+    return this.call("hello_client", name);
+  },
 };
 
-const server = new Server(dispatcher);
-
-for await (const listener of Deno.listen({
+for await (const conn of Deno.listen({
   hostname,
   port,
 })) {
-  console.log("Client has connected");
+  console.log("Session has connected");
+  const server = new Session(conn, dispatcher);
   server
-    .start(listener)
+    .listen()
     .then(() => console.log("Client has disconnected"))
     .catch((e) => console.error(e));
+  console.log(await server.call("hello_server", "Alice"));
+  console.log(await server.call("hello_client", "Alice"));
 }
