+++ mod.ts
@@ -1,5 +1,8 @@
-export { Client } from "./client.ts";
-export { Server } from "./server.ts";
-export type { Dispatcher } from "./server.ts";
+export { Session } from "./session.ts";
+export type { Dispatcher } from "./session.ts";
 export { encode, decode } from "./message.ts";
 export type { Request, Response, Notification, Message } from "./message.ts";
+
+// DEPRECATED
+export { Client } from "./client.ts";
+export { Server } from "./server.ts";
